+++ scripts/spells/spell-fire-shield-OnEffectDeletion.js
@@ -1,3 +1,0 @@
-Hooks.off("updateActor", game.FireShieldHookClientSpecificId);
-const effectSecondary = token.actor.effects.find(eff=>eff.getFlag('core','statusId').includes("Fire Shield"));
-await effectSecondary?.delete();